# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Buddies API
  version: 1.0.0
  description: |
    Dual-sided B2B reverse-auction marketplace API (consumer restaurants + suppliers).
    Auth: `Authorization: Bearer <supabase_jwt>` or dev `Bearer dev:<role>:<userId>`.
servers:
  - url: http://localhost:8000/v1
...
paths:
...
  /supplier/kyc:
    get:
      tags: [supplier]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: KYC profile }
    post:
      tags: [supplier]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Upsert KYC }
  /supplier/kyc/submit:
    post:
      tags: [supplier]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Submit for verification }
  /supplier/kyc/dev-verify:
    post:
      tags: [supplier]
      summary: Dev-only verify
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Verified }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...